Add tests directories, split tests into multiple files
Split tests in rooms and beers app (api_views, models, async etc. - to keep things organized). Add empty tests to rooms (async) and auth (api views).
