+++ backend/rooms/async_db.py
@@ -3,7 +3,7 @@ from datetime import timedelta
 from channels.db import database_sync_to_async
 from django.contrib.auth.models import User
 from django.core.exceptions import ObjectDoesNotExist
-from django.db.models.aggregates import Avg, Count
+from django.db.models.aggregates import Avg
 from django.db.models.fields import DecimalField
 from django.utils import timezone
