+++ backend/users/tests.py
@@ -1,3 +1,11 @@
 from django.test import TestCase
+from rest_framework.test import APIClient
 
-# Create your tests here.
+
+class AuthViewsTests(TestCase):
+    def setUp(self) -> None:
+        self.client = APIClient()
+
+    @classmethod
+    def setUpTestData(cls) -> None:
+        pass
